% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PEPBVS-deprecated.R, R/full_enumerationc.R,
%   R/mc3c.R
\name{PEPBVS-deprecated}
\alias{PEPBVS-deprecated}
\alias{full_enumeration_pep}
\alias{mc3_pep}
\title{Deprecated functions in package \pkg{PEPBVS}.}
\usage{
full_enumeration_pep(...)

mc3_pep(...)
}
\arguments{
\item{...}{Deprecated.}
}
\description{
The functions listed below are deprecated. 
Alternative functions with similar functionality are mentioned.
}
\details{
Defunct functions are: \code{full_enumeration_pep} and \code{mc3_pep}.
Please replace both with \code{\link{pep.lm}}
}
\keyword{internal}
